+++ cypress-shared/support/b2b/checkout.js
@@ -51,6 +51,7 @@ export function fillContactInfo() {
 
 export function verifyAddress(address) {
   it('Verify Auto fill Address in checkout', updateRetry(3), () => {
+    cy.storeOrderForm()
     cy.get('body').then(($shipping) => {
       if ($shipping.find(selectors.OpenShipping).length) {
         cy.get(selectors.OpenShipping, { timeout: 5000 }).click()